Move mosaic state up
The issue before was that it required the Background popup to be open.
Now the popup can be open or closed and the mosaics still load.
